pyproject.toml
[tool.poetry]
name = "elboto"
version = "0.0.1"
description = "Eloston's Discord bot"
authors = ["Eloston"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = "^3.7"
"discord.py" = "^1.6.0"
uvloop = "^0.15.0"
discord-ext-typed-commands = "^1.0.3"
"discord.py-stubs" = "^1.6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.800"
flake8 = "^3.8.4"
pytest = "^6.2.2"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
